# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: "PaperViz API"
  version: "0.1.0"
  description: |
    Research evidence understanding platform. Upload academic papers (PDF or pasted text), receive simplified-language summaries with re-visualized charts, and compare findings across papers.

    ## Authentication

    PaperViz uses session-based authentication via an httpOnly cookie (`session_token`, 7-day expiry).
    - **Signup/Login** creates a session cookie
    - **No API keys** — cookie-based authentication only
    - **Protected endpoints** require a valid session cookie

    ## Rate Limits

    | Endpoint | Limit | Window |
    |----------|-------|--------|
    | `POST /api/documents` | 1 request | 30 seconds |
    | All other endpoints | No limit | — |

    Rate limits are enforced per IP fingerprint (SHA256 of IP + User-Agent + Accept-Language).

    ## Submission Limits

    PaperViz is free and BYOK; there are no plans or paid tiers. The only
    submission cap is a per-device guard against automated bulk uploads on the
    unauthenticated creation endpoint.

    | Scope | Limit |
    |-------|-------|
    | Per IP fingerprint | 5 papers per month |

    Counts reset on the 1st of each month. Hitting the cap returns 429 with
    `{"error": "submission limit reached for this device", "papers_used": N, "limit": 5}`.

    ## Error Format

    All errors follow a consistent format:

    ```json
    { "error": "error_code" }
    ```

    | Code | HTTP Status | Meaning |
...
  contact:
...
servers:
...
security:
  - sessionCookie: []
paths:
  /api/auth/signup:
    post:
      operationId: signup
      summary: "Create a new user account"
      description: "Register with email and password. Creates a session cookie (httpOnly, 7-day expiry) on success."
      tags: ["Authentication"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
            examples:
              basic:
                summary: "Basic signup"
                value:
                  email: "user@example.com"
                  password: "securepass123"
      responses:
        "201":
          description: "Account created, session cookie set"
          headers:
            Set-Cookie:
              schema:
                type: string
                example: "session_token=abc123; Path=/; HttpOnly; SameSite=Lax; Max-Age=604800"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: "Invalid request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalid_email:
                  summary: "Invalid email"
                  value:
                    error: "invalid_email"
                password_too_short:
                  summary: "Password too short"
                  value:
                    error: "password_too_short"
        "409":
          description: "Email already taken"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/login:
    post:
      operationId: login
      summary: "Authenticate an existing user"
      description: "Login with email and password. Creates a session cookie (httpOnly, 7-day expiry) on success."
      tags: ["Authentication"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            examples:
              basic:
                summary: "Basic login"
                value:
                  email: "user@example.com"
                  password: "securepass123"
      responses:
        "200":
          description: "Login successful, session cookie set"
          headers:
            Set-Cookie:
              schema:
                type: string
                example: "session_token=abc123; Path=/; HttpOnly; SameSite=Lax; Max-Age=604800"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: "Invalid request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: "Invalid credentials"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/logout:
    post:
      operationId: logout
      summary: "End current session"
      description: "Clears the session cookie and invalidates the session."
      tags: ["Authentication"]
      security:
        - sessionCookie: []
      responses:
        "200":
          description: "Session ended"
          headers:
            Set-Cookie:
              schema:
                type: string
                example: "session_token=; Path=/; Max-Age=0"
          content:
            application/json:
              schema:
                type: object
  /api/auth/me:
    get:
      operationId: getMe
      summary: "Get current authenticated user"
      description: "Returns the user associated with the current session cookie."
      tags: ["Authentication"]
      security:
        - sessionCookie: []
      responses:
        "200":
          description: "Current user info"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: "Not authenticated"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/credentials:
    post:
      operationId: createCredential
      summary: "Store a model API key (BYOK)"
      description: |
        Encrypts the supplied key with AES-256-GCM and stores it against the
        authenticated user. PaperViz holds no model vendor account, so this key
        pays for the analysis.

        The response never contains the key. Only `key_hint`, the last four
        characters, is stored alongside it for display.
      tags: ["Credentials"]
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider, api_key]
              properties:
                provider:
                  type: string
                  enum: [gemini, anthropic, openai]
                  description: "Provider protocol is handled by github.com/zendev-sh/goai."
                model:
                  type: string
                  description: "Optional. Defaults to the provider's default model."
                api_key:
                  type: string
                  format: password
                  description: "The user's own provider key. Never logged, never returned."
      responses:
        "201":
          description: "Credential stored"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Credential"
        "400":
          description: "Unsupported provider or missing key"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: "Not authenticated"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listCredentials
      summary: "List stored credentials"
      description: "Returns metadata only. Ciphertext and plaintext keys are never included."
      tags: ["Credentials"]
      security:
        - sessionCookie: []
      responses:
        "200":
          description: "Credential list"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Credential"
        "401":
          description: "Not authenticated"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/credentials/{id}:
    delete:
      operationId: deleteCredential
      summary: "Delete a stored credential"
      description: "Scoped to the authenticated user. An unknown or unowned id is a no-op."
      tags: ["Credentials"]
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Deleted"
          content:
            application/json:
              schema:
                type: object
        "401":
          description: "Not authenticated"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/credentials/{id}/default:
    post:
      operationId: setDefaultCredential
      summary: "Promote a credential to default"
      description: |
        The credential used for analysis. Promoting one demotes the previous
        default; the database enforces at most one default per user.
      tags: ["Credentials"]
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Promoted"
          content:
            application/json:
              schema:
                type: object
        "404":
          description: "No such credential for this user"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/documents:
    post:
      operationId: createDocument
      summary: "Create a new document for analysis"
      description: |
        Upload a PDF (max 20 MiB) or paste text for analysis. The backend extracts text and chart images, then runs simplification, claim verification, chapter detection, and chart generation using the authenticated user's own model key.

        **Requires a configured credential.** There is no server-side model key, so a request from a user with no default credential fails with `missing_credential` (400) and no document is created.

        **Rate limit:** 1 request per 30 seconds per IP fingerprint.
        **Kill switch:** When `INGESTION_ENABLED=false`, this returns 503 with `Retry-After` while read, share, and library routes keep working.
        Processing runs in the background — poll `GET /api/documents/{id}` until status is `complete`.

        Accepts `multipart/form-data`.
      tags: ["Documents"]
      x-rateLimit:
        limit: 1
        window: 30
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateDocumentRequest"
            examples:
              pdf_upload:
                summary: "Upload PDF"
                encoding:
                  file:
                    contentType: application/pdf
              paste_text:
                summary: "Paste text"
                value:
                  text: "Your paper text here..."
                  reading_level: "simplified"
      responses:
        "201":
          description: "Document created, processing started"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateDocumentResponse"
        "400":
          description: "Invalid input"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                file_too_large:
                  summary: "File too large"
                  value:
                    error: "file_too_large"
                missing_input:
                  summary: "Missing input"
                  value:
                    error: "missing_input"
                invalid_file_type:
                  summary: "Invalid file type"
                  value:
                    error: "invalid_file_type"
                invalid_reading_level:
                  summary: "Invalid reading level"
                  value:
                    error: "invalid_reading_level"
        "429":
          description: "Rate limit exceeded or per-device submission limit reached"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                rate_limit:
                  summary: "Rate limit exceeded"
                  value:
                    error: "rate_limit_exceeded"
                submission_limit:
                  summary: "Per-device submission limit reached"
                  value:
                    error: "submission limit reached for this device"
                    papers_used: 5
                    limit: 5

    get:
      operationId: listDocuments
      summary: "List all documents for authenticated user"
      description: "Returns paginated list of documents owned by the current user. Requires authentication."
      tags: ["Documents"]
      security:
        - sessionCookie: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
          description: "Max results per page"
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: "Pagination offset"
      responses:
        "200":
          description: "Paginated document list"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentListResponse"
        "401":
          description: "Not authenticated"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/documents/compare:
    post:
      operationId: compareDocuments
      summary: "Compare multiple papers"
      description: "Compare 2-10 documents side-by-side. Extracts structured paper summaries, identifies agreement/disagreement, and generates cross-paper evidence claims."
      tags: ["Documents"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompareRequest"
            examples:
              two_papers:
                summary: "Compare two papers"
                value:
                  document_ids: ["doc_abc123", "doc_def456"]
      responses:
        "200":
          description: "Comparison complete"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comparison"
        "400":
          description: "Invalid document count (must be 2-10)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: "One or more documents not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: "Extraction or comparison failed"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/documents/stats:
    get:
      operationId: getDocumentStats
      summary: "Get document statistics"
      description: "Returns aggregate counts for the authenticated user's documents."
      tags: ["Documents"]
      security:
        - sessionCookie: []
      responses:
        "200":
          description: "Document statistics"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentStats"
        "401":
          description: "Not authenticated"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/documents/{id}:
    get:
      operationId: getDocument
      summary: "Get full document with analysis results"
      description: |
        Returns the complete document including simplified text, charts, chapters, claim diff, and evidence.
        Touches `last_accessed_at` to extend the 7-day expiry window.
      tags: ["Documents"]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "Document ID (e.g., doc_abc123)"
      responses:
        "200":
          description: "Full document with all analysis results"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: "Document not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateDocumentTitle
      summary: "Update document title"
      description: "Change the title of an existing document. Requires authentication and ownership."
      tags: ["Documents"]
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTitleRequest"
      responses:
        "200":
          description: "Title updated"
          content:
            application/json:
              schema:
                type: object
                properties:
                  title:
                    type: string
        "400":
          description: "Invalid input"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: "Not authenticated"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: "Document not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteDocument
      summary: "Delete a document"
      description: "Permanently delete a document and all associated data. Requires authentication and ownership."
      tags: ["Documents"]
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Document deleted"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: ["deleted"]
        "401":
          description: "Not authenticated"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: "Document not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/documents/{id}/save:
    put:
      operationId: toggleSaved
      summary: "Toggle saved/bookmark status"
      description: "Bookmark or unbookmark a document. Requires authentication and ownership."
      tags: ["Documents"]
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                saved:
                  type: boolean
              required: ["saved"]
      responses:
        "200":
          description: "Saved status updated"
          content:
            application/json:
              schema:
                type: object
                properties:
                  saved:
                    type: boolean
        "401":
          description: "Not authenticated"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: "Document not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/documents/{id}/charts/{chartId}/image:
    get:
      operationId: getChartImage
      summary: "Get chart as binary image"
      description: "Returns the raw image bytes for a chart captured via image fallback. Content-Type is auto-detected (image/png, image/jpeg, etc.)."
      tags: ["Documents"]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: chartId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Binary image data"
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
        "404":
          description: "Chart not found or no image"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/documents/{id}/charts/{chartId}/share:
    post:
      operationId: generateChartShareToken
      summary: "Generate share link for a chart"
      description: "Create a shareable token for a specific chart figure. Requires authentication and document ownership."
      tags: ["Sharing"]
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: chartId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Share token generated"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareTokenResponse"
        "401":
          description: "Not authenticated"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: "Not document owner"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: "Document or chart not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: revokeChartShareToken
      summary: "Revoke chart share link"
      description: "Remove the share token for a specific chart. Requires authentication and document ownership."
      tags: ["Sharing"]
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: chartId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: "Share token revoked"
        "401":
          description: "Not authenticated"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: "Document or chart not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/documents/{id}/share:
    post:
      operationId: generateDocShareToken
      summary: "Generate share link for entire document"
      description: "Create a shareable token for the full document. Requires authentication and document ownership."
      tags: ["Sharing"]
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Share URL generated"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocShareResponse"
        "401":
          description: "Not authenticated"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: "Document not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: revokeDocShareToken
      summary: "Revoke document share link"
      description: "Remove the share token for a document. Requires authentication and document ownership."
      tags: ["Sharing"]
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: "Share token revoked"
        "401":
          description: "Not authenticated"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: "Document not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /share-referrals:
    post:
      operationId: trackReferral
      summary: "Track share-to-analysis conversion"
      description: "Record a referral from a shared link. No authentication required."
      tags: ["Analytics"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ref:
                  type: string
                  description: "Share token value"
              required: ["ref"]
      responses:
        "204":
          description: "Referral tracked"
        "400":
          description: "Invalid ref token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_token
      description: "httpOnly session cookie set by `/api/auth/login` or `/api/auth/signup`. Expires after 7 days."
...